+++ app/DataTables/InvoicesTrashDataTable.php
@@ -4,10 +4,8 @@ namespace BT\DataTables;
 
 use BT\Modules\Invoices\Models\Invoice;
 use BT\Support\Statuses\InvoiceStatuses;
-use Yajra\DataTables\Services\DataTable;
-use Yajra\DataTables\Html\Column;
 
-class InvoicesTrashDataTable extends DataTable
+class InvoicesTrashDataTable extends InvoicesDataTable
 {
     /**
      * Build DataTable class.
@@ -64,6 +62,8 @@ class InvoicesTrashDataTable extends DataTable
     {
         return $this->builder()
             ->columns($this->getColumns())
+            ->removeColumn('amount.balance')
+            ->removeColumn('amount.total')
             ->ajax(['data' => 'function(d) { d.table = "invoices"; }'])
             ->orderBy(3, 'desc')
             ->lengthMenu([
@@ -72,54 +72,4 @@ class InvoicesTrashDataTable extends DataTable
             ]);
     }
 
-    /**
-     * Get columns.
-     *
-     * @return array
-     */
-    protected function getColumns()
-    {
-        return [
-            Column::make('id')
-                ->orderable(false)
-                ->searchable(false)
-                ->printable(false)
-                ->exportable(false)
-                ->className('bulk-record'),
-            Column::make('invoice_status_id')
-                ->title(trans('bt.status')),
-            Column::make('number')
-                ->title(trans('bt.invoice')),
-            Column::make('invoice_date')
-                ->title(trans('bt.date'))
-                ->data('formatted_invoice_date')
-                ->searchable(false),
-            Column::make('due_at')
-                ->title(trans('bt.due'))
-                ->data('formatted_due_at')
-                ->searchable(false),
-            Column::make('client_name')
-                ->name('client.name')
-                ->title(trans('bt.client'))
-                ->data('client.name'),
-            Column::make('summary')
-                ->title(trans('bt.summary'))
-                ->data('formatted_summary'),
-            Column::computed('action')
-                ->exportable(false)
-                ->printable(false)
-                ->width(80)
-                ->addClass('text-center'),
-        ];
-    }
-
-    /**
-     * Get filename for export.
-     *
-     * @return string
-     */
-    protected function filename()
-    {
-        return 'Invoices_' . date('YmdHis');
-    }
 }