+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - fix Vendor delete
 - added category, itemlookup and mail queue datatables
 - updated all datatables
+- add swalConfirm target (for sweetalert over iframe)
 
 ## 5.1.0
 - Added client shipping address fields (address_2,city_2,etc) accessible in templates